Partially revert c6dda68e6285fb6ae2a55ba45fad11fdb21ae285.
This sanity check makes sure that what we read into the buffer actually
is what we expect. The commit message claimed 'did not work when
messages from different sources mixed up', but that is a scenario that
isn't supposed to happen anyway: We should only get responses for things
we actually requested.